requirements.txt
python-dotenv
mysqlclient ~=1.3.0
Django ~=4.2
djangorestframework
djangorestframework-simplejwt
markdown
django-filter
coverage